+++ internal/machine/wait.go
@@ -6,6 +6,7 @@ import (
 	"fmt"
 	"net/http"
 	"strings"
+	"sync"
 	"time"
 
 	"github.com/jpillora/backoff"
@@ -13,6 +14,10 @@ import (
 	"github.com/superfly/fly-go/flaps"
 	"github.com/superfly/flyctl/internal/flapsutil"
 	"github.com/superfly/flyctl/internal/flyerr"
+	"github.com/superfly/flyctl/internal/statuslogger"
+	"github.com/superfly/flyctl/internal/tracing"
+	"go.opentelemetry.io/otel/attribute"
+	"go.opentelemetry.io/otel/trace"
 )
 
 func WaitForStartOrStop(ctx context.Context, machine *fly.Machine, action string, timeout time.Duration) error {
@@ -62,6 +67,71 @@ func WaitForStartOrStop(ctx context.Context, machine *fly.Machine, action string
 			}
 			time.Sleep(b.Duration())
 		}
+	}
+}
+
+// returns when the machine is in one of the possible states, or after passing the timeout threshold
+func WaitForAnyMachineState(ctx context.Context, mach *fly.Machine, possibleStates []string, timeout time.Duration, sl statuslogger.StatusLine) (string, error) {
+	ctx, span := tracing.GetTracer().Start(ctx, "wait_for_machine_state", trace.WithAttributes(
+		attribute.StringSlice("possible_states", possibleStates),
+	))
+	defer span.End()
+
+	ctx, cancel := context.WithTimeout(ctx, timeout)
+	defer cancel()
+
+	flapsClient := flapsutil.ClientFromContext(ctx)
+
+	var mutex sync.Mutex
+
+	var waitErr error
+	numCompleted := 0
+	var successfulState string
+
+	for _, state := range possibleStates {
+		state := state
+		go func() {
+			err := flapsClient.Wait(ctx, mach, state, timeout)
+			mutex.Lock()
+			defer func() {
+				numCompleted += 1
+				mutex.Unlock()
+			}()
+
+			if successfulState != "" {
+				return
+			}
+
+			if sl != nil {
+				sl.LogStatus(statuslogger.StatusRunning, fmt.Sprintf("Machine %s reached %s state", mach.ID, state))
+			}
+
+			if err != nil {
+				waitErr = err
+			} else {
+				successfulState = state
+			}
+		}()
+	}
+
+	// TODO(billy): i'm sure we can use channels here
+	for {
+		mutex.Lock()
+		if successfulState != "" || numCompleted == len(possibleStates) {
+			defer mutex.Unlock()
+			if successfulState != "" {
+				span.SetAttributes(attribute.String("state", successfulState))
+			}
+
+			if waitErr != nil {
+				span.RecordError(waitErr)
+			}
+
+			return successfulState, waitErr
+		}
+		mutex.Unlock()
+
+		time.Sleep(1 * time.Second)
 	}
 }
